[REF] iot_drivers: simplify ctypes drivers
`CtypesTerminalDriver` class is the parent class of both
Six and Worldline (L/W) drivers.

Some code parts were duplucated between these drivers, so
we moved duplicated parts in the parent class.
